backend/requirements.txt
flask
flask_sqlalchemy
flask_limiter
flask_caching
python-dotenv
openai
